+++ R/Read_Results.R
@@ -9,7 +9,7 @@
 #' @importFrom pdftools pdf_text
 #' @importFrom readr read_delim
 #'
-#' @param file a .pdf or .html file (could be a url) where containing swimming
+#' @param file a pdf, url or Hytek .hy3 file containing swimming
 #'   results.  Must be formatted in a "normal" fashion - see vignette
 #' @param node a CSS node where html results are stored.  Required for html
 #'   results.  Default is "pre", which nearly always works.
@@ -33,41 +33,32 @@ Read_Results <- function(file, node = "pre") {
   #### begin actual function ####
   if (stringr::str_detect(file, "\\.pdf$|\\.aspx$") == TRUE) {
     ### PDF ###
-    results <- pdftools::pdf_text(file)
-    as_lines <- stringr::str_extract_all(results, "\n.*")
-    as_lines_list_2 <- unlist(as_lines, recursive = FALSE)
-    as_lines_list_2 <- read_results_flag(as_lines_list_2)
 
-    return(as_lines_list_2)
+    results <- read_pdf(file)
 
-  } else if (stringr::str_detect(file, "\\.htm") == TRUE) {
-    ### HTML ###
-    if (is.character(node) == FALSE) {
-      stop(" Please supply a value for node")
-    } else {
-      webpage <- xml2::read_html(file)
-      html <- rvest::html_nodes(webpage, node)
-      results <- rvest::html_text(html)
-      as_lines <- stringr::str_extract_all(results, "\n.*")
-      as_lines_list_2 <- unlist(as_lines, recursive = FALSE)
-      as_lines_list_2 <- read_results_flag(as_lines_list_2)
-
-      return(as_lines_list_2)
-
+    if(all("error" %in% class(results), stringr::str_detect(file, "\\.aspx$")) ==TRUE){
+      results <- file %>%
+        read_htm(node_helper = node)
     }
-  } else if (stringr::str_detect(file, "\\.hy3") == TRUE) {
-    ### hy3 ###
-    ### add automatic unzipping?
-    results <-
-      readr::read_delim(file, delim = "\\s\\2", col_names = FALSE)
-    as_lines_list_2 <- unlist(results, recursive = FALSE)
-    as_lines_list_2 <- read_results_flag(as_lines_list_2)
-    row_numbs <- seq(1, length(as_lines_list_2), 1)
-    as_lines_list_2 <- paste(as_lines_list_2, row_numbs, sep = "  ")
-    return(as_lines_list_2)
-  } else {
-    stop("Please supply a valid .html, .hy3 or .pdf document")
-  }
+
+    } else if (stringr::str_detect(file, "\\.htm|\\.aspx$") == TRUE) {
+        ### HTML ###
+      results <- file %>%
+        read_htm(node_helper = node)
+      } else if (stringr::str_detect(file, "\\.hy3") == TRUE) {
+        ### hy3 ###
+        ### add automatic unzipping?
+        results <- file %>%
+          read_hy3()
+
+      }
+
+      if("error" %in% class(results) == TRUE){
+        stop("Please supply a valid .html, .hy3 or .pdf document")
+      }
+
+
+  return(results)
 }
 
 
@@ -89,5 +80,81 @@ read_results <- Read_Results
 
 read_results_flag <- function(x){
   x <- c("read_results_flag", x)
+
+}
+
+#' Read in pdf files of swimming results
+#'
+#' Based on pdftools, this function can be temperamental
+#'
+#' @param x a .pdf or .aspx location containing swimming results.  Must be
+#'   formatted in a "normal" fashion - see vignette
+#' @return returns a list of results, with "read_results_flag" added as the
+#'   first element of the list
+#'
+
+read_pdf <- function(x){
+  suppressMessages(
+    results <-
+      tryCatch(
+        pdftools::pdf_text(x),
+        error = function(e)
+          e
+      )
+  )
+
+  if ("error" %in% class(results) == FALSE) {
+    results <- stringr::str_extract_all(results, "\n.*")
+    results <- unlist(results, recursive = FALSE)
+    results <- read_results_flag(results)
+  }
+
+  return(results)
+}
+
+#' Read in html files of swimming results
+#'
+#' @param x an .html, .htm or .aspx location containing swimming results.  Must be
+#'   formatted in a "normal" fashion - see vignette
+#' @param node_helper receives \code{node} from \code{read_results}
+#' @return returns a list of results, with "read_results_flag" added as the
+#'   first element of the list
+#'
+
+read_htm <- function(x, node_helper) {
+  if (is.character(node_helper) == FALSE) {
+    stop(" Please supply a value for node")
+  }
+
+  webpage <- xml2::read_html(x)
+  html <- rvest::html_nodes(webpage, node_helper)
+  results <- rvest::html_text(html)
+  results <- stringr::str_extract_all(results, "\n.*")
+  results <- unlist(results, recursive = FALSE)
+  results <- read_results_flag(results)
+
+  return(results)
+
+}
+
+
+#' Read in hy3 files of swimming results
+#'
+#' @param x an unzipped hy3 file containing swimming results.  Must be
+#'   formatted in a "normal" fashion - see vignette
+#' @return returns a list of results, with "read_results_flag" added as the
+#'   first element of the list
+#'
+
+read_hy3 <- function(x) {
+  results <-
+    readr::read_delim(file, delim = "\\s\\2", col_names = FALSE)
+  results <- unlist(results, recursive = FALSE)
+  results <- read_results_flag(results)
+  row_numbs <- seq(1, length(results), 1)
+  results <-
+    paste(results, row_numbs, sep = "  ")
+
+  return(results)
 
 }